requirements.txt
#------
# Conda
#------

matplotlib
pytorch torchvision torchaudio cudatoolkit=10.2 -c pytorch
scipy
tqdm